fix fetchers comments
