+++ package.json
@@ -26,13 +26,14 @@
     "fastify": "^4.0.0",
     "fastify-cli": "^5.7.1",
     "fastify-plugin": "^4.0.0",
-    "node-fetch": "^2.6.6",
+    "node-fetch": "2",
     "pg": "^8.11.1",
     "typeorm": "^0.3.17"
   },
   "devDependencies": {
     "@types/config": "^3.3.0",
     "@types/node": "^18.0.0",
+    "@types/node-fetch": "^2.6.4",
     "@types/tap": "^15.0.5",
     "@typescript-eslint/eslint-plugin": "^5.61.0",
     "@typescript-eslint/parser": "^5.61.0",
