+++ test/routes/root.test.ts
@@ -3,7 +3,6 @@ import { build } from '../helper';
 
 test('default root route', async (t) => {
   const app = await build(t);
-  //root
   const res = await app.inject({
     url: '/',
   });
@@ -23,16 +22,12 @@ test('default root route', async (t) => {
     }
   };
 
-  // Parse the response payload
   const responseData1 = JSON.parse(res1.payload);
-
-  // Check if the response data has the same structure as the expected structure
   t.match(responseData1, expectedDataStructure1);
 
   //fetcher2
   const res2 = await app.inject({
     url: '/fetcher2',
   });
-  // Parse the response payload
   t.match(JSON.parse(res2.payload), { price: Number });
 });
